fix web lint for admin redirect helper
- Move the admin post-login redirect helper into a standalone util module so App.tsx stays component-only for Fast Refresh.
- Add the helper to the web source sets used by eslint and the app build so Bazel can resolve it during linting.
- Keep the redirect-target unit coverage pointed at the extracted helper.
